+++ examples/storybook/src/shared/utils/mockBookingsData.ts
@@ -6,6 +6,8 @@ export default [
       'https://www.wix.app/bookings/54e7b326-49ec-434f-a7cc-c2ad1eaec049/service/03ef6546-985d-4739-951e-ca897375b967/details?d=https://sapirs0.wixsite.com/mysite/bookings-checkout/cupcake-master',
     name: 'CUPCAKE MASTER',
     imageSrc: 'https://static.wixstatic.com/media/78ccafc88f7fce3e55bd7a33f166704b.jpg',
+    imageWidth: 1036,
+    imageHeight: 984,
     description: 90,
     durations: 90,
     html:
@@ -19,6 +21,8 @@ export default [
     description: 30,
     durations: 30,
     imageSrc: 'https://static.wixstatic.com/media/6fb7a82098b749b8bd98773dc3bf6aea.jpg',
+    imageWidth: 5616,
+    imageHeight: 3744,
     html:
       '<a class="vertical-embed-link" href="https://www.wix.app/bookings/54e7b326-49ec-434f-a7cc-c2ad1eaec049/service/05ea21a5-6bae-41d6-bb83-a67ce7b7eaae/details?d=https://sapirs0.wixsite.com/mysite/bookings-checkout/baking" target="_blank"><div class="vertical-embed-card vertical-embed-ltr vertical-embed-cardLayout"><div style="background-image:url(https://static.wixstatic.com/media/6fb7a82098b749b8bd98773dc3bf6aea.jpg)" class="vertical-embed-imageLayout vertical-embed-image"></div><div class="vertical-embed-content"><div><div class="vertical-embed-title">BAKING</div></div><div class="vertical-embed-button"><div class="vertical-embed-buttonText">Reservar ahora</div></div></div></div></a>',
   },
@@ -30,6 +34,8 @@ export default [
     description: 100,
     durations: 100,
     imageSrc: 'https://static.wixstatic.com/media/11062b_443aaea08ad643aab6ee734a217abf2b~mv2.jpg',
+    imageWidth: 4500,
+    imageHeight: 3000,
     html:
       '<a class="vertical-embed-link" href="https://www.wix.app/bookings/54e7b326-49ec-434f-a7cc-c2ad1eaec049/service/910974b8-093b-422e-92ef-4ac66dece7d9/details?d=https://sapirs0.wixsite.com/mysite/bookings-checkout/how-to-make-artisan-sourdough-bread" target="_blank"><div class="vertical-embed-card vertical-embed-ltr vertical-embed-cardLayout"><div style="background-image:url(https://static.wixstatic.com/media/11062b_443aaea08ad643aab6ee734a217abf2b~mv2.jpg)" class="vertical-embed-imageLayout vertical-embed-image"></div><div class="vertical-embed-content"><div><div class="vertical-embed-title">How to make artisan sourdough bread???</div></div><div class="vertical-embed-button"><div class="vertical-embed-buttonText">Reservar ahora</div></div></div></div></a>',
   },
@@ -49,6 +55,8 @@ export default [
       'https://www.wix.app/bookings/54e7b326-49ec-434f-a7cc-c2ad1eaec049/service/ce5e4de0-5428-4a39-be82-3126fe53f42e/details?d=https://sapirs0.wixsite.com/mysite/bookings-checkout/frosting',
     name: 'FROSTING',
     imageSrc: 'https://static.wixstatic.com/media/87141f9c9a074b11b38181fb303ba00a.jpg',
+    imageWidth: 3600,
+    imageHeight: 2697,
     html:
       '<a class="vertical-embed-link" href="https://www.wix.app/bookings/54e7b326-49ec-434f-a7cc-c2ad1eaec049/service/ce5e4de0-5428-4a39-be82-3126fe53f42e/details?d=https://sapirs0.wixsite.com/mysite/bookings-checkout/frosting" target="_blank"><div class="vertical-embed-card vertical-embed-ltr vertical-embed-cardLayout"><div style="background-image:url(https://static.wixstatic.com/media/87141f9c9a074b11b38181fb303ba00a.jpg)" class="vertical-embed-imageLayout vertical-embed-image"></div><div class="vertical-embed-content"><div><div class="vertical-embed-title">FROSTING</div></div><div class="vertical-embed-button"><div class="vertical-embed-buttonText">Reservar ahora</div></div></div></div></a>',
   },