+++ packages/plugin-social-polls/web/src/helpers.ts
@@ -1,5 +1,4 @@
-import imageClientAPI from 'image-client-api/dist/imageClientSDK';
-
+import { getScaleImageSrc } from 'wix-rich-content-common/libs/imageUtils';
 import { BACKGROUND_TYPE } from './defaults';
 
 export function generateId() {
@@ -32,32 +31,11 @@ export function getBackgroundString(
         return null;
       }
       // eslint-disable-next-line no-case-declarations
-      const imageSrc = getImageSrc(background, width, height);
+      const imageSrc = getScaleImageSrc(background, width, height);
       return `url('${imageSrc}') center / cover`;
 
     case BACKGROUND_TYPE.COLOR:
     default:
       return background;
-  }
-}
-
-export function getMediaId(src: string) {
-  try {
-    const [, mediaId] = /media\/([^/]+)/.exec(src) as string[];
-    return mediaId;
-  } catch (error) {
-    return src;
-  }
-}
-
-export function getImageSrc(src: string, width: number, height: number) {
-  const mediaId = getMediaId(src);
-
-  try {
-    return imageClientAPI.getScaleToFillImageURL(mediaId, null, null, width, height, {
-      quality: 90,
-    });
-  } catch (error) {
-    return src;
   }
 }