+++ packages/plugin-vertical-embed/web/src/components/Card.js
@@ -1,14 +1,34 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import CardContent from './CardContent';
 import styles from '../../statics/styles/widget.scss';
 import classNames from 'classnames';
+import { getScaleImageSrc } from 'wix-rich-content-common/libs/imageUtils';
+
+const getImageTargetHeight = (width, height, targetWidth) => {
+  const proportion = height / width;
+  const targetHeight = proportion * targetWidth;
+  return targetHeight;
+};
 
 const Card = props => {
-  const { url, imageSrc, content, direction } = props;
+  const { url, content, direction, imageWidth, imageHeight } = props;
+  const [imageSrc, setImageSrc] = useState();
+
+  const setRef = ref => {
+    if (ref && !imageSrc) {
+      const targetWidth = ref.getBoundingClientRect()?.width;
+      if (imageWidth && imageHeight) {
+        const targetHeight = getImageTargetHeight(imageWidth, imageHeight, targetWidth);
+        setImageSrc(getScaleImageSrc(props.imageSrc, targetWidth, targetHeight));
+      } else {
+        setImageSrc(props.imageSrc);
+      }
+    }
+  };
   return (
     // eslint-disable-next-line react/jsx-no-target-blank
-    <a className={styles.link} href={url} target="_blank">
+    <a className={styles.link} href={url} target="_blank" ref={setRef}>
       <div
         style={{ direction }}
         className={classNames(styles[direction], styles.container, styles.cardLayout)}
@@ -27,6 +47,8 @@ Card.propTypes = {
   content: PropTypes.object,
   direction: PropTypes.string.isRequired,
   url: PropTypes.string.isRequired,
+  imageWidth: PropTypes.string,
+  imageHeight: PropTypes.string,
 };
 
 export default Card;