+++ packages/plugin-vertical-embed/web/src/utils.ts
@@ -19,33 +19,39 @@ export const convertDuration = (durationInMinutes, t) => {
   return durationInHours;
 };
 
+const getBaseData = ({ imageSrc, pageUrl, imageWidth, imageHeight }) => ({
+  url: pageUrl,
+  imageSrc,
+  imageWidth,
+  imageHeight,
+});
+
 const getBookingData = (data, t) => {
-  const { name, imageSrc, pageUrl, durations } = data;
+  const { name, durations } = data;
   const content = {
     title: name,
     info: { leftSubtitle: durations && convertDuration(durations, t) },
     buttonText: t('VerticalEmbed_Bookings_Button'),
   };
-  return { url: pageUrl, imageSrc, content };
+  return { content, ...getBaseData(data) };
 };
 
 const getEventData = (data, t) => {
-  const { name, imageSrc, scheduling, pageUrl, location } = data;
+  const { name, scheduling, location } = data;
   const content = {
     title: name,
     info: { leftSubtitle: scheduling, rightSubtitle: location },
     buttonText: t('VerticalEmbed_Events_Button'),
   };
-  return { url: pageUrl, imageSrc, content };
+  return { content, ...getBaseData(data) };
 };
 
 const getProductData = (data, t) => {
-  const { name, imageSrc, pageUrl } = data;
   const content = {
-    title: name,
+    title: data.name,
     buttonText: t('VerticalEmbed_Products_Button'),
   };
-  return { url: pageUrl, imageSrc, content };
+  return { content, ...getBaseData(data) };
 };
 
 export const dataTypeMapper = {